Changement de nom du dossier models Ajout d' un ressource temporaire de boite skinnée et affichage (temporaire aussi) dans le projet Test Ajout des Vector4f Chargement des données de skinning (pas encore contenu dans le VB) Noms plus logiques pour les fonctions Get...Matrix de Camera

+++ src/src/Engine/Core/Inputs/InputManager.cpp
@@ -180,8 +180,8 @@ LRESULT CALLBACK InputManager::EventsCallback( HWND hWnd, UINT uMsg, WPARAM wPar
 
 			if( pointMouse.x!=posX || pointMouse.y != posY )
 			{
-				m_MouseOffset.x = pointMouse.x - posX;
-				m_MouseOffset.y = pointMouse.y - posY;
+				m_MouseOffset.x = (FLOAT)pointMouse.x - (FLOAT)posX;
+				m_MouseOffset.y = (FLOAT)pointMouse.y - (FLOAT)posY;
 
 				SetCursorPos( posX, posY );
 			}
